refactor: default extension RecentProjects into integrated extensions folder
