+++ src/base-config/keyboard.json
@@ -337,5 +337,10 @@
     ],
     "help.support": [
         "Shift-F1"
+    ],
+    "recentProjects.toggle": [
+        {
+            "key": "Ctrl-Alt-R"
+        }
     ]
 }
