+++ src/extensions/default/DefaultExtensions.json
@@ -21,7 +21,6 @@
     "Phoenix-prettier",
     "PrefsCodeHints",
     "QuickView",
-    "RecentProjects",
     "SVGCodeHints",
     "UrlCodeHints",
     "HealthData",
