--- src/extensions/default/RecentProjects/keyboard.json
@@ -1,7 +1,0 @@
-{
-    "recentProjects": [
-        {
-            "key": "Ctrl-Alt-R"
-        }
-    ]
-}
