--- src/extensions/default/RecentProjects/package.json
@@ -1,6 +1,0 @@
-{
-  "name": "RecentProjects",
-  "version": "1.0.0",
-  "dependencies": {
-  }
-}
