--- src/extensions/default/RecentProjects/requirejs-config.json
@@ -1,1 +1,0 @@
-{}
