+++ src/extensionsIntegrated/RecentProjects/htmlContent/projects-menu.html
