+++ src/extensionsIntegrated/loader.js
@@ -39,4 +39,5 @@ define(function (require, exports, module) {
     require("./NoDistractions/main");
     require("./Phoenix-live-preview/main");
     require("./NavigationAndHistory/main");
+    require("./RecentProjects/main");
 });
