+++ src/styles/Extn-RecentProjects.less
@@ -15,7 +15,7 @@
     width: 7px;
     height: 5px;
     margin-left: 4px;
-    background-image: url("down-arrow.svg");
+    background-image: url("images/down-arrow.svg");
     background-repeat: no-repeat;
     position: relative;
     top: 0;
@@ -60,7 +60,7 @@
 }
 
 #project-dropdown-toggle:hover .dropdown-arrow {
-    background-image: url("down-arrow.svg");
+    background-image: url("images/down-arrow.svg");
 }
 
 #project-dropdown.dropdown-menu:focus {
