+++ src/styles/brackets.less
@@ -38,7 +38,9 @@
  * right place
  */
 @import "brackets_shared.less";
-@import "extensions/NavigationAndHistory.less";
+// Integrated extensions import
+@import "Extn-NavigationAndHistory.less";
+@import "Extn-RecentProjects.less";
 
 /* Overall layout */
