+++ test/UnitTestSuite.js
@@ -113,7 +113,8 @@ define(function (require, exports, module) {
     // Integrated extension tests
     require("spec/Extn-InAppNotifications-integ-test");
     require("spec/Extn-RemoteFileAdapter-integ-test");
-    require("spec/Extn-NavigationAndHistory-integ-tests");
+    require("spec/Extn-NavigationAndHistory-integ-test");
+    require("spec/Extn-RecentProjects-integ-test");
     // Node Tests
     require("spec/NodeConnection-test");
     // todo TEST_MODERN
